Create README.md
Liver Diseases are a leading cause of death among people. There are a number of liver diseases, the most significant of them being liver tumors. Liver tumors are an abnormal growth of liver cells on or in the liver. This project uses the fundamentals of Image Processing and Convolutional Neural Networks to detect liver diseases.

Segmentation of liver lesions in Computed Tomography images can be used to assess the intensity of the disease, plan treatments, predict and monitor the clinical response. Convolutional Neural Networks have been used for liver disease segmentation, which has been modeled mathematically to identify abnormalities in a healthy liver. The Computed Tomography scan-based lesion type definition defines the diagnosis and therapeutic strategies. It demands a great deal of experience, expertise and resources.

A Deep Learning approach is used to examine abdominal Computed Tomography images of the liver and distinguish between healthy and unhealthy livers. The proposed method includes the successful extraction of features right from formation combined with residual and pre-trained weights. Feature maps have been consistent with the original image volume pixel and the features represent the most relevant imaging criteria for every class. This deep learning system highlights various portions of the decision-making process of a pre-trained Deep Neural Network, through an analysis of inner layers and the description of features that lead to predictions. The Internet of Things system is equipped with a Global System for Mobile Communications module as well as a WiFi module. The WiFi module continually updates the status of a patient to a server and the Global System for Mobile Communications module automatically sends out an alert message to the patient’s emergency contact when a tumor has been detected.